requirements.txt
streamlit
google-generativeai
crewai
langchain-google-genai
duckduckgo-search
fpdf2
python-dotenv
